@@ -3,10 +3,21 @@ name = "supplier"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.12"
+requires-python = ">=3.14"
 dependencies = [
-    "confluent-kafka>=2.6.0",
-    "opentelemetry-distro>=0.48b0",
-    "opentelemetry-exporter-otlp>=1.27.0",
-    "pydantic>=2.9.2",
+    "common-models",
+    "confluent-kafka>=2.12.2",
+    "opentelemetry-distro>=0.59b0",
+    "opentelemetry-exporter-otlp>=1.38.0",
+    "pydantic>=2.12.4",
 ]
+
+[dependency-groups]
+dev = [
+    "ruff",
+    "black",
+    "pyright",
+]
+
+[tool.uv.sources]
+common-models = { path = "../common-models" }
