@@ -2,13 +2,11 @@ import os
 import random
 import json
 import logging
-import os
-import random
 import time
 
 from confluent_kafka import Producer
 
-from common.common.models import Order, WoodType
+from common_models.models import Order, WoodType
 
 # Configure the logger with environment variable
 log_level = os.environ.get("LOG_LEVEL", "INFO").upper()
@@ -55,9 +53,7 @@ if __name__ == "__main__":
         # Simulate random error for observability testing
         # The error rate is controlled by the ERROR_RATE environment variable (default: 0.1)
         if random.random() < ERROR_RATE:
-            logger.error(
-                "failed to send order (Kafka/network failure)"
-            )
+            logger.error("failed to send order (Kafka/network failure)")
             time.sleep(int(interval_seconds))
             continue
 
