@@ -11,7 +11,7 @@ from ..crud import (
     get_orders_by_status,
     update_order_status,
 )
-from ..database import get_db
+from ..database import SessionLocal
 from ..models import OrderStatus
 
 # Configure Flask/Werkzeug logging to show HTTP errors as ERROR level
@@ -82,16 +82,22 @@ def create_order_route():
     if random.random() < ERROR_RATE:
         return jsonify({"error": "Simulated DB insertion error"}), 500
     data = request.json
-    db = next(get_db())
-    order_data = schemas.OrderCreate(**data)
+    if not data:
+        return jsonify({"error": "Missing request body"}), 400
+    
+    db = SessionLocal()
     try:
+        order_data = schemas.OrderCreate(**data)
         new_order = create_order(db=db, order=order_data)
-    except Exception:
+        return jsonify(new_order.to_dict()), 201
+    except Exception as e:
+        current_app.logger.exception("Unexpected error during order creation")
         return (
             jsonify({"error": "Unexpected error during order creation"}),
             500,
         )
-    return jsonify(new_order.to_dict()), 201
+    finally:
+        db.close()
 
 
 @orders_bp.route("/", methods=["GET"])
@@ -137,15 +143,19 @@ def read_orders_route():
         )
     skip = int(request.args.get("skip", 0))
     limit = int(request.args.get("limit", 10))
-    db = next(get_db())
+    
+    db = SessionLocal()
     try:
         orders = get_orders(db=db, skip=skip, limit=limit)
-    except Exception:
+        return jsonify([order.to_dict() for order in orders])
+    except Exception as e:
+        current_app.logger.exception("Unexpected error during order list")
         return (
             jsonify({"error": "Unexpected error during order list"}),
             500,
         )
-    return jsonify([order.to_dict() for order in orders])
+    finally:
+        db.close()
 
 
 @orders_bp.route("/status/<status>", methods=["GET"])
@@ -195,17 +205,21 @@ def read_orders_by_status_route(status):
         )
     skip = int(request.args.get("skip", 0))
     limit = int(request.args.get("limit", 10))
-    db = next(get_db())
+    
+    db = SessionLocal()
     try:
         orders = get_orders_by_status(
             db=db, order_status=OrderStatus(status), skip=skip, limit=limit
         )
-    except Exception:
+        return jsonify([order.to_dict() for order in orders])
+    except Exception as e:
+        current_app.logger.exception("Unexpected error during order list by status")
         return (
             jsonify({"error": "Unexpected error during order list by status"}),
             500,
         )
-    return jsonify([order.to_dict() for order in orders])
+    finally:
+        db.close()
 
 
 @orders_bp.route("/<int:order_id>", methods=["GET"])
@@ -243,17 +257,21 @@ def read_order_route(order_id):
             jsonify({"error": "Simulated API error during order read"}),
             502,
         )
-    db = next(get_db())
+    
+    db = SessionLocal()
     try:
         db_order = get_order(db=db, order_id=order_id)
-    except Exception:
+        if db_order is None:
+            return jsonify({"error": "Order not found"}), 404
+        return jsonify(db_order.to_dict())
+    except Exception as e:
+        current_app.logger.exception("Unexpected error during order read")
         return (
             jsonify({"error": "Unexpected error during order read"}),
             500,
         )
-    if db_order is None:
-        return jsonify({"error": "Order not found"}), 404
-    return jsonify(db_order.to_dict())
+    finally:
+        db.close()
 
 
 @orders_bp.route("/<int:order_id>", methods=["PUT"])
@@ -301,7 +319,7 @@ def update_order_status_route(order_id):
             502,
         )
     data = request.json
-    if "order_status" not in data:
+    if not data or "order_status" not in data:
         return jsonify({"error": "Missing 'order_status' in request body"}), 400
 
     try:
@@ -309,24 +327,25 @@ def update_order_status_route(order_id):
     except ValueError:
         return jsonify({"error": "Invalid order status"}), 400
 
-    db = next(get_db())
+    db = SessionLocal()
     try:
         order = get_order(db, order_id=order_id)
-    except Exception:
-        return (
-            jsonify({"error": "Unexpected error during order fetch for update"}),
-            500,
-        )
-    if order is None:
-        return jsonify({"error": "Order not found"}), 404
+        if order is None:
+            return jsonify({"error": "Order not found"}), 404
 
-    try:
         updated_order = update_order_status(
             db, order_id=order_id, order_status=order_status
         )
-    except Exception:
+        
+        if updated_order is None:
+            return jsonify({"error": "Order not found"}), 404
+
+        return jsonify(updated_order.to_dict()), 200
+    except Exception as e:
+        current_app.logger.exception("Unexpected error during order status update")
         return (
             jsonify({"error": "Unexpected error during order status update"}),
             500,
         )
-    return jsonify(updated_order.to_dict()), 200
+    finally:
+        db.close()
