@@ -1,10 +1,9 @@
 import random
 import os
 import time
-import json
 import logging
 import requests
-from common.common.models import OrderStatus
+from common_models.models import OrderStatus
 
 # Configure logging with environment variable
 log_level = os.environ.get("LOG_LEVEL", "INFO").upper()
@@ -49,12 +48,30 @@ def decrease_stock(order):
 
 
 def update_order_status(order_id, status):
-    payload = {"order_status": status}
-    response = requests.put(
-        f"{API_URL_ORDERS_UPDATE}/{order_id}", headers=HEADERS_JSON, json=payload
-    )
-    response.raise_for_status()
-    return response.json()
+    payload = {"order_status": status.value if hasattr(status, "value") else status}
+    try:
+        response = requests.put(
+            f"{API_URL_ORDERS_UPDATE}/{order_id}", headers=HEADERS_JSON, json=payload
+        )
+        response.raise_for_status()
+        return response.json()
+    except requests.RequestException as e:
+        # If the response is available, include its body for richer logs
+        resp_text = None
+        if hasattr(e, "response") and e.response is not None:
+            try:
+                resp_text = e.response.text
+            except Exception:
+                resp_text = "<unable to read response body>"
+        logger.error(
+            "Failed to update order %s status to %s: %s; response=%s",
+            order_id,
+            payload,
+            e,
+            resp_text,
+        )
+        # Re-raise so callers can decide how to handle it
+        raise
 
 
 def process_registered_order():
@@ -72,9 +89,7 @@ def process_registered_order():
             # The error rate is controlled by the ERROR_RATE environment variable (default: 0.1)
             if random.random() < ERROR_RATE:
                 # Simulate an external API or DB failure
-                raise Exception(
-                    "external API or DB failure during order processing"
-                )
+                raise Exception("external API or DB failure during order processing")
 
             logger.info("Processing order: %s", order)
             decrease_stock(order)
@@ -83,8 +98,23 @@ def process_registered_order():
             logger.info("Order status updated to READY for order: %s", order)
         except Exception as e:
             logger.error(f"Failed to process order {order['id']}: {e}")
-            update_order_status(order["id"], OrderStatus.BLOCKED)
-            logger.info("Order status updated to BLOCKED for order: %s", order)
+            # Attempt to mark the order as BLOCKED, but don't let a failed
+            # status update crash the whole worker. Log failures and continue.
+            try:
+                update_order_status(order["id"], OrderStatus.BLOCKED)
+                logger.info("Order status updated to BLOCKED for order: %s", order)
+            except requests.RequestException as re:
+                logger.error(
+                    "Could not mark order %s as BLOCKED: %s. Continuing.",
+                    order["id"],
+                    re,
+                )
+            except Exception as re:
+                logger.error(
+                    "Unexpected error while marking order %s as BLOCKED: %s. Continuing.",
+                    order["id"],
+                    re,
+                )
 
 
 if __name__ == "__main__":
