@@ -1,6 +1,7 @@
+# pyright: reportMissingImports=false
 from sqlalchemy import Column, Enum, Integer
 
-from common.common.models import WoodType
+from common_models.models import WoodType
 
 from .database import db
 
