@@ -1,6 +1,7 @@
+# pyright: reportMissingImports=false
 from pydantic import BaseModel, Field
 
-from common.common.models import WoodType
+from common_models.models import WoodType
 
 
 class StockBase(BaseModel):
